Add WebAssembly/Emscripten port for browser play
- Add Makefile.emscripten with SDL2 port, ASYNCIFY, and IDBFS support
- Add sdl1_compat.c to bridge SDL1 APIs (SetVideoMode, DisplayFormat, etc.) to SDL2
- Adapt main.cpp, tablestorage.cpp, opensave.cpp for Emscripten (IDBFS persistence, resource paths)
- Fix SDL2 text input handling in widgets.cpp (SDL_TEXTINPUT events)
- Fix narrowing conversion error in font.cpp for clang/emcc
- Add shell.html for browser embedding
- Add docs/ with built WASM files for GitHub Pages deployment

+++ font.cpp
@@ -72,7 +72,7 @@ void Font::draw(SDL_Surface *s, int x, int y, int r, int g, int b,
         SDL_BlitSurface(surface, &src, s, &dst);
         SDL_FreeSurface(surface);
     }
-    SDL_Color color = { r, g, b, 0 };
+    SDL_Color color = { (Uint8)r, (Uint8)g, (Uint8)b, 0 };
     SDL_Surface *surface = TTF_RenderUNICODE_Blended(font, str, color);
     SDL_Rect src = { 0, 0, surface->w, surface->h };
     SDL_Rect dst = { x, y, surface->w, surface->h };
